# gateway (copy)/docs/swagger.yaml
basePath: /
definitions:
...
  genproto.NotificationCreate:
    properties:
      Message:
        type: string
      RecieverId:
        type: string
      SenderId:
        type: string
    type: object
  genproto.NotificationGet:
    properties:
      CreatedAt:
        type: string
      Id:
        type: string
      Message:
        type: string
      SenderId:
        type: string
      Status:
        type: string
      UserId:
        type: string
    type: object
  genproto.NotificationList:
    properties:
      Limit:
        type: integer
      Notifications:
        items:
          $ref: '#/definitions/genproto.NotificationGet'
        type: array
      Offset:
        type: integer
      TotalCount:
        type: integer
    type: object
  genproto.NotificationMessage:
    properties:
      Message:
        type: string
      SenderId:
        type: string
      TargetGroup:
        type: string
    type: object
  genproto.NotificationUpt:
    properties:
      Message:
        type: string
      Status:
        type: string
    type: object
...
info:
  contact: {}
  description: API for Instant Delivery resources
  title: Finance Tracker API Documentation
  version: "1.0"
paths:
...
  /notifications/{notification_id}:
    delete:
      consumes:
      - application/json
      description: Delete a notification by its ID
      parameters:
      - description: Notification ID
        in: path
        name: notification_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Notification deleted successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Notification not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Notification
      tags:
      - Notification
    get:
      consumes:
      - application/json
      description: Retrieve a notification by its ID
      parameters:
      - description: Notification ID
        in: path
        name: notification_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Notification details
          schema:
            $ref: '#/definitions/genproto.NotificationGet'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Notification not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Notification
      tags:
      - Notification
    put:
      consumes:
      - application/json
      description: Update an existing notification
      parameters:
      - description: Notification ID
        in: path
        name: notification_id
        required: true
        type: string
      - description: Notification data
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/genproto.NotificationUpt'
      produces:
      - application/json
      responses:
        "200":
          description: Notification updated successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Notification
      tags:
      - Notification
  /notifications/create:
    post:
      consumes:
      - application/json
      description: Create a new notification
      parameters:
      - description: Notification data
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/genproto.NotificationCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Notification created successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Notification
      tags:
      - Notification
  /notifications/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of notifications
      parameters:
      - description: Receiver ID
        in: query
        name: reciever_id
        type: string
      - description: Notification Status
        in: query
        name: status
        type: string
      - description: Sender ID
        in: query
        name: sender_id
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of notifications
          schema:
            $ref: '#/definitions/genproto.NotificationList'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List Notifications
      tags:
      - Notification
  /notifications/notify-all:
    post:
      consumes:
      - application/json
      description: Send a notification to all users in the target group
      parameters:
      - description: Notification data
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/genproto.NotificationMessage'
      produces:
      - application/json
      responses:
        "200":
          description: Notification sent successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Notify All
      tags:
      - Notification
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"